requirements.txt
Django>=4.0
psycopg2>=2.9.0
celery>=5.3.6
redis>=5.0.1
django-environ
gunicorn
djangorestframework